applayer: add missing error handling on loadMasterList
Coverity CCID 559398

+++ pteid-mw-pt/_src/eidmw/common/Util.cpp
@@ -473,7 +473,7 @@ EIDMW_CMN_API int vfprintf_s(FILE *stream, const char *format, va_list argptr) {
 
 #endif
 
-EIDMW_CMN_API size_t read_binary_file(const char *filename, unsigned char **outBuffer) {
+EIDMW_CMN_API long read_binary_file(const char *filename, unsigned char **outBuffer) {
 	FILE *file = fopen(filename, "rb");
 	if (file == NULL) {
 		perror("Failed to open file");
